pyproject.toml
[project]
name = "solarwatch"
version = "0.1.0"
description = "European PV Digital Experience Monitoring System"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
dependencies = [
    "sqlalchemy>=2.0,<3.0",
    "pydantic>=2.0,<3.0",
    "pydantic-settings>=2.0,<3.0",
    "pyyaml>=6.0,<7.0",
    "google-play-scraper>=1.2,<2.0",
    "requests>=2.28,<3.0",
    "google-genai>=1.0,<2.0",
    "openai>=1.0,<2.0",
    "streamlit>=1.30,<2.0",
    "plotly>=5.18,<6.0",
    "altair>=5.0,<6.0",
    "pandas>=2.1,<3.0",
    "rich>=13.0,<14.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4,<9.0",
    "pytest-asyncio>=0.23,<1.0",
    "pytest-cov>=4.1,<6.0",
    "ruff>=0.2,<1.0",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
target-version = "py39"
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
